third_party/WebKit/LayoutTests/editing/execCommand/selectAll.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  '<span contenteditable id="test">|foo bar<p>baz\nabc</p></span>',
  'selectAll',
  '<span contenteditable id="test">^foo bar<p>baz\nabc|</p></span>'),
  'selectAll');
</script>
